+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changed
 
 - Flag `-p, --project` exposed to allow querying sessions by project
+- Introduction of `-a, --all` flag to override default timespan settings
 
 # [2.0.3] - 2021-01-31